+++ Model/DateTimeEncoder.py
@@ -1,0 +1,13 @@
+import json
+
+from datetime import datetime as dt
+
+
+class DateTimeEncoder(json.JSONEncoder):
+    def default(self, o):
+        if isinstance(o, dt):
+            return o.isoformat()
+
+        if isinstance(o, bytes):
+            return list(o)
+        return json.JSONEncoder.default(self, o)
