+++ View/LoginCodeWindow.py
@@ -1,0 +1,10 @@
+# import asyncio
+#
+# from kivy.properties import ObjectProperty
+# from kivy.uix.screenmanager import Screen
+# from telethon import errors
+#
+# from View.LoginWindow import LoginWindow
+# from main import tlAgent, sm
+
+
