+++ View/Shared.py
@@ -1,0 +1,6 @@
+from View.WindowManager import WindowManager
+from tlagent import TLAgent
+
+sm = WindowManager()
+my_users = []
+tlAgent = TLAgent
