+++ View/WindowManager.py
@@ -1,0 +1,5 @@
+from kivy.uix.screenmanager import ScreenManager
+
+class WindowManager(ScreenManager):
+    pass
+
